moving unix socket into samples
