--- samples/unix_socket_broker.py
@@ -1,15 +1,0 @@
-from amqtt.broker import Broker
-from amqtt.contexts import BrokerConfig, ListenerConfig
-
-
-async def main():
-
-    cfg = BrokerConfig(
-        listeners={
-            'default': ListenerConfig(
-                ListenerType.External
-            )
-        }
-    )
-
-    b = Broker()
